requirements.txt
### NSFR
lark-parser
scikit-learn
setuptools>=58.5.3
tensorboard
numpy>=1.18.5
tqdm>=4.41.0
matplotlib>=3.2.2
opencv-python==4.5.5.64
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
seaborn
pandas
rtpt